requirements.txt
numpy
Pillow
certifi==2018.4.16
chardet==3.0.4
flickrapi==2.4.0
idna==2.7
oauthlib==2.1.0
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.23
